+++ src/dune_rules/coq_rules.ml
@@ -191,7 +191,7 @@ let select_native_mode ~sctx ~dir (buildable : Coq_stanza.Buildable.t) =
     else if buildable.coq_lang_version < (0, 7) then Memo.return Coq_mode.VoOnly
     else
       let* coqc = resolve_program sctx ~dir ~loc:buildable.loc "coqc" in
-      let+ config = Coq_config.make ~bin:(Action.Prog.ok_exn coqc) in
+      let+ config = Coq_config.make ~coqc in
       match Coq_config.by_name config "coq_native_compiler_default" with
       | Some (`String "yes") | Some (`String "ondemand") -> Coq_mode.Native
       | _ -> Coq_mode.VoOnly)