+++ test/blackbox-tests/test-cases/melange/merlin.t
@@ -16,7 +16,7 @@
 
   $ touch bar.ml $lib.ml
   $ dune build @check
-  $ dune ocaml-merlin --dump-config="$PWD" | grep -i "$lib"
+  $ dune ocaml merlin dump-config "$PWD" | grep -i "$lib"
   Foo
     $TESTCASE_ROOT/_build/default/.foo.objs/melange)
      Foo__
@@ -43,7 +43,7 @@ All 3 entries (Foo, Foo__ and Bar) contain a ppx directive
 
   $ touch main.ml
   $ dune build @check
-  $ dune ocaml-merlin --dump-config="$PWD" | grep -i "$target"
+  $ dune ocaml merlin dump-config "$PWD" | grep -i "$target"
     $TESTCASE_ROOT/_build/default/.output.mobjs/melange)
 
 The melange.emit entry contains a ppx directive